Add .scan() method for databanks
A method to get all items of a type. Useful for e.g. backing up a
databank. Also for generating roll-up data (count) if you haven't been
storing it before.

Most drivers already implement something like it internally
for search without an index, so it seemed safe to surface it.